+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # main
 
+* Updated gemspec file. ([#30](https://github.com/petalmd/rubocop-petal/pull/30))
 * Added cop `RSpec/JsonParseResponseBody and RSpec/JsonResponse` ([#27](https://github.com/petalmd/rubocop-petal/pull/27))
 
 # v0.7.0