+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # main
 
+# v0.8.0 (2022-02-21)
+
 * Added cop `Performance/Snif` ([#31](https://github.com/petalmd/rubocop-petal/pull/31))
 * Updated gemspec file. ([#30](https://github.com/petalmd/rubocop-petal/pull/30))
 * Added cop `RSpec/JsonParseResponseBody and RSpec/JsonResponse` ([#27](https://github.com/petalmd/rubocop-petal/pull/27))